Provider service modifier addition

+++ Furniture-Moving-Project/src/main/java/SWP301/Furniture_Moving_Project/dto/ProviderDTO.java
@@ -3,20 +3,20 @@ package SWP301.Furniture_Moving_Project.dto;
 import java.math.BigDecimal;
 
 public class ProviderDTO {
-    private Integer providerId;
+    private Long providerId;
     private String companyName;
     private BigDecimal rating;
 
     public ProviderDTO() {
     }
-    public ProviderDTO(Integer providerId, String companyName, BigDecimal rating) {
+    public ProviderDTO(Long providerId, String companyName, BigDecimal rating) {
         this.providerId = providerId;
         this.companyName = companyName;
         this.rating = rating;
     }
 
-    public Integer getProviderId() { return providerId; }
-    public void setProviderId(Integer providerId) { this.providerId = providerId; }
+    public Long getProviderId() { return providerId; }
+    public void setProviderId(Long providerId) { this.providerId = providerId; }
     public String getCompanyName() { return companyName; }
     public void setCompanyName(String companyName) { this.companyName = companyName; }
     public BigDecimal getRating() { return rating; }
